% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answered_questions.R
\name{answered_questions}
\alias{answered_questions}
\alias{legco_answered_questions}
\title{Answered Questions in LegCo}
\usage{
answered_questions(speaker_id = NULL, member_id = NULL, type = "all",
  lang = "en", from = "1900-01-01", to = Sys.Date(), floor = FALSE,
  n = 50, verbose = TRUE)

legco_answered_questions(speaker_id = NULL, member_id = NULL,
  type = "all", lang = "en", from = "1900-01-01", to = Sys.Date(),
  floor = FALSE, n = 50, verbose = TRUE)
}
\arguments{
\item{speaker_id}{The Speaker ID, or a vector of IDs, as specified in the
output from the function `legco::speakers()`. Defaults to `NULL`.}

\item{member_id}{The member ID, or a vector of IDs, as specified in the output of the
function `legco::member()`. Defaults to `NULL`.}

\item{type}{The type of question. `'oral'` returns oral questions.
`'written'` returns written questions. `'all'` returns all questions.
Defaults to `'all'`.}

\item{lang}{The language of hansard files to search from. `'en'` returns the
English version. `'zh'` returns the Traditional Chinese version. Defaults
to `'en'`.}

\item{from}{Only fetch results from hansards of meetings on or after this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results from hansards of meetings on or before this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to the current system date.}

\item{floor}{Whether to fetch results from the floor version of the hansard
files. The floor version is the first presented version of hansard file in
the original language delivered by the speakers in LegCo. If `'TRUE'`, the
language option is ignored. Defaults to `FALSE`.}

\item{n}{The number of entry to fetch. Defaults to `50`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch full text and answered questions put to the government by specific
LegCo member(s).
}
