% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division_date.R
\name{division_date}
\alias{division_date}
\alias{legco_member_division_date}
\title{Division by Date}
\usage{
division_date(target_date = NULL, committee_id = NULL,
  meet_id = NULL, verbose = TRUE)

legco_member_division_date(target_date = NULL, committee_id = NULL,
  meet_id = NULL, verbose = TRUE)
}
\arguments{
\item{target_date}{The date and time when the division(s) was/were conducted. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`.}

\item{committee_id}{The id of a committee, or a vector of IDs. If `NULL`,
returns voting records of all committees. Note that not all committees have
its voting records available. Defaults to `NULL`.}

\item{meet_id}{The id of a meeting. If `NULL`, returns voting records of all
committee meetings. Note that not all committees have its voting records
available. Defaults to `NULL`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch division(s) conducted on a specified date.
}
