% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_voting_record.R
\name{member_voting_record}
\alias{member_voting_record}
\title{Voting Records of a LegCo Member}
\usage{
member_voting_record(speaker_id = NULL, member_id = NULL,
  from = "1900-01-01", to = Sys.Date(), n = 50, verbose = TRUE)
}
\arguments{
\item{speaker_id}{The Speaker ID, or a vector of IDs. If `NULL`, returns all
LegCo members. Defaults to `NULL`.}

\item{member_id}{The Member ID, or a vector of IDs. Defaults to `NULL`.}

\item{from}{Only fetch results from hansards of meetings on or after this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results from hansards of meetings on or before this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to the current system date.}

\item{n}{The number of entry to fetch. Defaults to `50`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch voting records of a specific LegCo member.
}
