% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_committee.R
\name{search_committee}
\alias{search_committee}
\alias{legco_search_committee}
\title{Search Committee}
\usage{
search_committee(search_term, verbose = TRUE)

legco_search_committee(search_term, verbose = TRUE)
}
\arguments{
\item{search_term}{The search term. Accept full or partial Chinese or English
name of a committee.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Search committee by full or partial name of committee.
}
