% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_division.R
\name{search_division}
\alias{search_division}
\alias{legco_search_division}
\title{Search Division}
\usage{
search_division(search_date = NULL, committee_id = NULL,
  slot_id = NULL, index = TRUE, verbose = TRUE, ...)

legco_search_division(search_date = NULL, committee_id = NULL,
  slot_id = NULL, index = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{search_date}{The date and time when the division(s) was/were conducted.
Accepts character values in `'YYYY-MM-DD'` format, and objects of class
`Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced
to a date with `as.Date()`.}

\item{committee_id}{The id of a committee, or a vector of IDs.}

\item{slot_id}{The id of a meeting slot. If `NULL`, returns voting records of
all committee meetings.}

\item{index}{If `TRUE`, returns an index of votes. If `FALSE`, returns the
voting record of individual members. Defaults to `TRUE`.}

\item{verbose}{Defaults to `TRUE`.}

\item{...}{Parameters supported by `legco::voting_record()`.}
}
\description{
Fetch division(s) (votes) conducted by a given comittee(s) or on a given date
or during a given meeting slot. Note that only voting records from the
Council, House Committee, Finance Committee and its subcomittees are
available.
}
