% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_meeting.R
\name{search_meeting}
\alias{search_meeting}
\alias{legco_search_meeting}
\title{Search meeting}
\usage{
search_meeting(committee_id, from = "1990-01-01", to = Sys.Date(),
  verbose = TRUE)

legco_search_meeting(committee_id, from = "1990-01-01",
  to = Sys.Date(), verbose = TRUE)
}
\arguments{
\item{committee_id}{The id of a committee, or a vector of IDs.}

\item{from}{Only fetch meetings on or after this date. Accepts character
values in `'YYYY-MM-DD'` format, and objects of class `Date`, `POSIXt`,
`POSIXct`, `POSIXlt` or anything else that can be coerced to a date with
`as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only meetings on or before this date. Accepts character values in
`'YYYY-MM-DD'` format, and objects of class `Date`, `POSIXt`, `POSIXct`,
`POSIXlt` or anything else that can be coerced to a date with `as.Date()`.
Defaults to the current system date.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetches meetings of a given committee.
}
