% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_voting_record.R
\name{member_voting_record}
\alias{member_voting_record}
\alias{legco_member_voting_record}
\title{Voting Records of a LegCo Member}
\usage{
member_voting_record(speaker_id = NULL, member_id = NULL,
  committee_id = NULL, slot_id = NULL, from = "1990-01-01T00:00:00",
  to = Sys.time(), n = 10000, verbose = TRUE, ...)

legco_member_voting_record(speaker_id = NULL, member_id = NULL,
  committee_id = NULL, slot_id = NULL, from = "1990-01-01T00:00:00",
  to = Sys.time(), n = 10000, verbose = TRUE, ...)
}
\arguments{
\item{speaker_id}{The Speaker ID, or a vector of IDs. If `NULL`, returns all
LegCo members. Defaults to `NULL`.}

\item{member_id}{The Member ID, or a vector of IDs. If `NULL`, returns all
LegCo members. Defaults to `NULL`.}

\item{committee_id}{The id of a committee, or a vector of IDs. If `NULL`,
returns voting records of all committees. Note that not all committees have
its voting records available. Defaults to `NULL`.}

\item{slot_id}{The id of a meeting slot. If `NULL`, returns voting records of all
committee meetings. Note that not all committees have its voting records
available. Defaults to `NULL`.}

\item{from}{Only fetch votes conducted at or after this time. Accepts
character values in `'YYYY-MM-DDTHH:MM:SS'` format, and objects of class
`Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced
to a time with `as.POSIXlt()`. Defaults to `'1900-01-01T00:00:00'`.}

\item{to}{Only fetch votes conducted at or before this time. Accepts
character values in `'YYYY-MM-DDTHH:MM:SS'` format, and objects of class
`Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced
to a time with `as.POSIXlt()`. Defaults to system time.}

\item{n}{The number of entry to fetch. Defaults to `10000`.}

\item{verbose}{Defaults to `TRUE`.}

\item{...}{Parameters supported by `legco::voting_record()`.}
}
\description{
Fetch voting records of LegCo member(s).
}
