% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_committee.R
\name{search_committee}
\alias{search_committee}
\alias{legco_search_committee}
\title{Search Committee}
\usage{
search_committee(search_term, term_id = NULL, exact = FALSE,
  verbose = TRUE)

legco_search_committee(search_term, term_id = NULL, exact = FALSE,
  verbose = TRUE)
}
\arguments{
\item{search_term}{The search term. Accept full or partial Chinese or English
name of a committee.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, searches
committees from all terms. Defaults to `NULL`.}

\item{exact}{Whether to look for exact match only. If `TRUE`, return only
exact matches. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Search committee by full or partial name of committee.
}
